new module for health related classes
